+++ src/components/animate-counter/index.ts
@@ -1,0 +1,2 @@
+export { MPAnimateCounter } from './MPAnimateCounter';
+export type { MPAnimateCounterProps } from './MPAnimateCounter';