+++ src/index.ts
@@ -11,6 +11,7 @@ export * from './components/accordion';
 export * from './components/alert';
 export * from './components/animate-appear';
 export * from './components/animate-blink';
+export * from './components/animate-counter';
 export * from './components/animate-fade';
 export * from './components/animate-float';
 export * from './components/animate-grow';